backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
openai>=1.0.0
python-dotenv==1.0.0
chromadb==0.4.18
scikit-learn==1.3.2